ledger & parkstay: move oracle interface table to ledger payments
