+++ ledger/payments/models.py
@@ -7,6 +7,8 @@ from ledger.payments.invoice.models import Invoice, InvoiceBPAY
 from ledger.payments.bpoint.models import BpointTransaction, BpointToken
 from ledger.payments.cash.models import CashTransaction
 
+# Oracle Integration
+# ======================================
 class OracleParser(models.Model):
     inserted = models.DateTimeField(auto_now_add=True)
     date_parsed = models.DateField()
@@ -18,3 +20,15 @@ class OracleParserInvoice(models.Model):
     parser = models.ForeignKey(OracleParser)
     reference = models.CharField(max_length=50)
     details = JSONField() 
+
+class OracleInterface(models.Model):
+    receipt_number = models.IntegerField()
+    receipt_date = models.DateField()
+    activity_name = models.CharField(max_length=50)
+    amount = models.DecimalField(max_digits=8, decimal_places=2)
+    customer_name = models.CharField(max_length=128)
+    description = models.TextField()
+    comments = models.TextField()
+    status = models.CharField(max_length=15)
+    line_item = models.TextField(blank=True,null=True)
+    status_date = models.DateField()
