+++ parkstay/admin.py
@@ -112,9 +112,6 @@ class ClosureReason(ReasonAdmin):
 class OpenReason(ReasonAdmin):
     pass
 
-@admin.register(models.OracleInterface)
-class OracleInterfaceAdmin(admin.ModelAdmin):
-    pass
 
 admin.site.register(models.Rate)
 admin.site.register(models.Region)
